FEAT: Added aez converter and formated aez file
The CLi tool to convert the official AEZ file, using the exclusion list
file (as a geojson). Once converted, the filtered S2 tile IDs associated
with each aez is available for the workplan to use as input, as well as
a curated list of season indicators and the flag for L8 optical
activation.